pyproject.toml
[project]
name = "marketing-ai"
version = "0.1.0"
description = "Marketing Lead Enrichment Agent System"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "simple-salesforce>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "msal>=1.31.0",
    "python-multipart>=0.0.17",
    "jinja2>=3.1.0",
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
